requirements.txt
Django==2.1.4
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-dotenv==1.4.2
django-versatileimagefield==1.10
gunicorn
